requirements.txt
Flask
requests
cryptography
python-telegram-bot>=20.0